feat: modularize header component
