+++ src/app/components/Header.tsx
@@ -1,24 +1,13 @@
 'use client';
-import Image from 'next/image';
-import { useEffect, useState } from 'react';
 import { FaSearch } from 'react-icons/fa';
+import { UserProfile } from './UserProfile';
+import { NavLinks } from './NavLinks';
+import { Logo } from './Logo';
+import { useScroll } from '../hooks/useScroll';
 
 export default function Header() {
-  const [isScrolled, setIsScrolled] = useState(false);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      // This is where we set the state of isScrolled
-      setIsScrolled(window.scrollY > 0);
-    };
-
-    // This is where we listen for the scroll event
-    window.addEventListener('scroll', handleScroll);
-
-    // This is a cleanup function that React will run when the component unmounts
-    // this is important to prevent memory leaks
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, [isScrolled]);
+  // This is where we call our custom hook
+  const isScrolled = useScroll();
 
   return (
     <header
@@ -27,21 +16,16 @@ export default function Header() {
       }`}
     >
       <div className='flex items-center space-x-8'>
-        <Image src='/netflix-logo.png' width={100} height={100} alt='Netflix' />
-        <ul className='hidden md:flex md:space-x-4'>
-          <li>Home</li>
-          <li>TV Shows</li>
-          <li>Movies</li>
-          <li>Latest</li>
-        </ul>
+        <Logo />
+        <NavLinks />
       </div>
 
       <div className='flex items-center space-x-2 md:space-x-4'>
         <button className='hidden md:inline'>
           <FaSearch />
         </button>
-        <span className='hidden md:inline'>Kids</span>
-        <Image src='/netflix-avatar-profile.png' alt='Profile Avatar' height={30} width={30} className='rounded-sm' />
+
+        <UserProfile />
       </div>
     </header>
   );
