+++ src/app/components/NavLinks.tsx
@@ -1,0 +1,18 @@
+export const NavLinks = () => (
+  <nav>
+    <ul className='hidden md:flex md:space-x-4'>
+      <li>
+        <a href='#'>Home</a>
+      </li>
+      <li>
+        <a href='#'>TV Shows</a>
+      </li>
+      <li>
+        <a href='#'>Movies</a>
+      </li>
+      <li>
+        <a href='#'>Latest</a>
+      </li>
+    </ul>
+  </nav>
+);
